:recycle: Move to util
