+++ src/swiftui/util/type_guards.ts
